requirements.txt
streamlit
feedparser
openai
python-dateutil